requirements.txt
alembic==1.12.1
async-timeout==4.0.3
asyncpg==0.29.0
colorama==0.4.6
greenlet==3.0.1
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
TgCrypto==1.2.5
typing_extensions==4.8.0
win32-setctime==1.1.0
